Use one integration target target
